api/pyproject.toml
[tool.poetry]
name = "financier"
version = "0.1.0"
description = ""
authors = ["Ariel Catli <dev.arielcatli@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aws-cdk-lib = "2.148.0"
constructs = ">=10.0.0,<11.0.0"
python-dotenv = "^1.0.1"
boto3 = "^1.35.29"
boto3-stubs = "^1.35.29"
pynamodb = "^6.0.1"
mypy = "^1.11.2"
black = "^24.8.0"
isort = "^5.13.2"
pydantic = "^2.9.2"
pytest = "^8.3.3"
pytest-sugar = "^1.0.0"
moto = {extras = ["all"], version = "^5.0.16"}
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
